requirements.txt
flask
pandas
openai
tqdm
pinecone-client
